Installer: Split installer.py into steps
PEP8 fixes
