+++ installer/steps/c_nltk.py
@@ -1,0 +1,6 @@
+from helper import *
+
+
+section("Download additional data (Dictionary)")
+
+shell('python -m nltk.downloader -d jarviscli/data/nltk wordnet', True).should_not_fail()
